PragGOToDocuments/Template/template/template-athulV/requirements.txt
openai
python-dotenv
psycopg2-binary
django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
